♻️ Consolidate test fixtures and mocks into entity barrel exports
