+++ src/entities/comment/index.ts
@@ -1,5 +1,11 @@
+export {
+  CommentFixtures,
+  CommentRepositoryMocks,
+  type MockCommentRepository,
+} from "./__tests__";
 export { Comment, CommentFactory, type CommentRepository } from "./core";
 export { COMMENT_QUERY_KEY } from "./infrastructure/api";
 export type { CommentDto } from "./infrastructure/dto";
 export { CommentApiRepository } from "./infrastructure/repository";
 export { CommentMapper } from "./mapper";
+export { type CommentEntity } from "./types";
