+++ src/entities/post/index.ts
@@ -1,5 +1,11 @@
+export {
+  PostFixtures,
+  PostRepositoryMocks,
+  type MockPostRepository,
+} from "./__tests__";
 export { Post, PostFactory, type PostRepository } from "./core";
 export { POST_QUERY_KEYS } from "./infrastructure/api/post-query-key.api";
 export type { PostDto } from "./infrastructure/dto";
 export { PostApiRepository } from "./infrastructure/repository";
 export { PostMapper } from "./mapper";
+export { type PostEntity } from "./types";
