+++ src/entities/user/index.ts
@@ -1,6 +1,12 @@
+export {
+  UserFixtures,
+  UserRepositoryMocks,
+  type MockUserRepository,
+} from "./__tests__";
 export { User, UserFactory, type UserRepository } from "./core";
 export { USER_QUERY_KEYS } from "./infrastructure/api";
 export type { UserProfileDto } from "./infrastructure/dto";
 export { UserApiRepository } from "./infrastructure/repository";
 export { UserMapper } from "./mapper";
+export { type UserEntity } from "./types";
 export { UserAvatar, UserIdentifier } from "./ui/identifier";
